LPS-69111 SF: Tests: Move IndexName and Index to top level classes
https://issues.liferay.com/browse/LPS-69111
